split schema.go in files + add unit test
