+++ pkg/model/manifest_friendly_err_test.go
@@ -80,7 +80,7 @@ yaml: some random error
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := newManifestFriendlyError(tt.input)
+			err := NewManifestFriendlyError(tt.input)
 			assert.Equal(t, tt.expected, err.Error())
 		})
 	}
