+++ pkg/model/schema_test.go
@@ -1,10 +1,10 @@
-// Copyright 2023 The Okteto Authors
+// Copyright 2023|2024 The Okteto Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
+// 
 // http://www.apache.org/licenses/LICENSE-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -220,7 +220,7 @@ func Test_getStructKeys(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			keys := getStructKeys(tt.input)
+			keys := GetStructKeys(tt.input)
 			for k, v := range tt.expected {
 				assert.ElementsMatch(t, v, keys[k])
 			}
